test: split optimized integration tests
